+++ src/database/server/models/__tests__/nextauth.test.ts
@@ -5,7 +5,7 @@ import type {
   AdapterUser,
   VerificationToken,
 } from '@auth/core/adapters';
-import { eq } from 'drizzle-orm';
+import { eq } from 'drizzle-orm/expressions';
 import { afterAll, afterEach, beforeAll, beforeEach, describe, expect, it, vi } from 'vitest';
 
 import {