+++ src/database/server/models/plugin.ts
@@ -19,6 +19,10 @@ export class PluginModel {
     const [result] = await this.db
       .insert(installedPlugins)
       .values({ ...params, createdAt: new Date(), updatedAt: new Date(), userId: this.userId })
+      .onConflictDoUpdate({
+        set: { ...params, updatedAt: new Date() },
+        target: [installedPlugins.identifier, installedPlugins.userId],
+      })
       .returning();
 
     return result;
@@ -56,10 +60,10 @@ export class PluginModel {
     });
   };
 
-  async update(id: string, value: Partial<InstalledPluginItem>) {
+  update = async (id: string, value: Partial<InstalledPluginItem>) => {
     return this.db
       .update(installedPlugins)
       .set({ ...value, updatedAt: new Date() })
       .where(and(eq(installedPlugins.identifier, id), eq(installedPlugins.userId, this.userId)));
-  }
+  };
 }