+++ src/store/session/slices/sessionGroup/action.test.ts
@@ -40,7 +40,9 @@ describe('createSessionGroupSlice', () => {
 
   describe('clearSessionGroups', () => {
     it('should clear session groups and refresh sessions', async () => {
-      vi.spyOn(sessionService, 'removeSessionGroups');
+      const spyOn = vi
+        .spyOn(sessionService, 'removeSessionGroups')
+        .mockResolvedValueOnce(undefined);
       const spyOnRefreshSessions = vi.spyOn(useSessionStore.getState(), 'refreshSessions');
 
       const { result } = renderHook(() => useSessionStore());
@@ -49,7 +51,7 @@ describe('createSessionGroupSlice', () => {
         await result.current.clearSessionGroups();
       });
 
-      expect(sessionService.removeSessionGroups).toHaveBeenCalled();
+      expect(spyOn).toHaveBeenCalled();
       expect(spyOnRefreshSessions).toHaveBeenCalled();
     });
   });
@@ -57,7 +59,7 @@ describe('createSessionGroupSlice', () => {
   describe('removeSessionGroup', () => {
     it('should remove a session group and refresh sessions', async () => {
       const mockId = 'mock-id';
-      vi.spyOn(sessionService, 'removeSessionGroup');
+      vi.spyOn(sessionService, 'removeSessionGroup').mockResolvedValueOnce(undefined);
       const spyOnRefreshSessions = vi.spyOn(useSessionStore.getState(), 'refreshSessions');
 
       const { result } = renderHook(() => useSessionStore());
@@ -75,7 +77,7 @@ describe('createSessionGroupSlice', () => {
     it('should update a session group id and refresh sessions', async () => {
       const mockSessionId = 'session-id';
       const mockGroupId = 'group-id';
-      vi.spyOn(sessionService, 'updateSession');
+      vi.spyOn(sessionService, 'updateSession').mockResolvedValueOnce(undefined);
       const spyOnRefreshSessions = vi.spyOn(useSessionStore.getState(), 'refreshSessions');
 
       const { result } = renderHook(() => useSessionStore());
@@ -96,7 +98,7 @@ describe('createSessionGroupSlice', () => {
       const mockId = 'mock-id';
       const mockName = 'New Name';
       const spyOnRefreshSessions = vi.spyOn(useSessionStore.getState(), 'refreshSessions');
-      vi.spyOn(sessionService, 'updateSessionGroup');
+      vi.spyOn(sessionService, 'updateSessionGroup').mockResolvedValueOnce(undefined);
 
       const { result } = renderHook(() => useSessionStore());
 
@@ -115,7 +117,7 @@ describe('createSessionGroupSlice', () => {
         { id: 'id1', sort: 0 },
         { id: 'id2', sort: 1 },
       ];
-      vi.spyOn(sessionService, 'updateSessionGroupOrder');
+      vi.spyOn(sessionService, 'updateSessionGroupOrder').mockResolvedValueOnce(undefined);
       const spyOnRefreshSessions = vi.spyOn(useSessionStore.getState(), 'refreshSessions');
 
       const { result } = renderHook(() => useSessionStore());